+++ src/script/generateBundles.js
@@ -1,20 +1,6 @@
 #!/usr/bin/env node
 const debug = require('debug');
 const fs = require('fs');
-const SvgLoader = require('svg-inline-loader');
-const IsomorphicTools = require('webpack-isomorphic-tools');
-const hook = require('node-hook');
-
-hook.hook('.scss', () => {});
-hook.hook('.svg', (source) => {
-  const markup = SvgLoader.getExtractedSVG(source, { removeSVGTagAttrs: false });
-  return 'module.exports = ' + JSON.stringify(markup);
-});
-const isoConfig = {
-  assets: { images: { extensions: ['png', 'jpg', 'gif', 'ico'] } }
-};
-const isoTools = new IsomorphicTools(isoConfig);
-
 const log = debug('toga:generateBundles');
 const getConfig = require('../app/config')();
 const runWebpack = require('./webpack');
@@ -64,21 +50,21 @@ const requireComponents = () => {
   return staticLocals;
 };
 
+let stats;
 module.exports = Promise.resolve()
   .then(() => {
-    return isoTools.server(process.cwd());
+    return runWebpack({ minify: !dev, entry, rules, commonsChunkName: vendor.componentName, staticComponents  });
+  })
+  .then((assetsBundlingStats) => {
+    stats = assetsBundlingStats;
+    const universalHelper = require('./ssr-helper');
+    return universalHelper();
   })
   .then(() => {
-    const staticLocals = requireComponents(staticComponents);
-    const promises = [];
-    promises.push(runWebpack({ minify: !dev, entry, rules, commonsChunkName: vendor.componentName, staticComponents, staticLocals  }));
-    if (staticComponents) {
-      promises.push(runWebpack({ entry:  { static: './src/script/webpack/staticRender.js' }, staticComponents, staticLocals }));
-    }
-    return Promise.all(promises);
+    const staticLocals = (staticComponents) ? requireComponents(staticComponents) : { };
+    return runWebpack({ entry:  { static: './src/script/webpack/staticRender.js' }, staticComponents, staticLocals });
   })
-  .then(([ assetsBundlingStats ]) => {
-    const stats = assetsBundlingStats;
+  .then(() => {
     fs.writeFile(process.cwd() + '/dist/webpack-components-stats.json', JSON.stringify(stats.toJson({chunkModules: true})), function(err) {
       if(err) {
         return log(err);
